LoopStudioWeb/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
Werkzeug==3.0.3
